requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
feedparser
openai
requests
python-dotenv
